gh-108303: Move `test_future` into its own test_future_stmt subdir (#109368)
(cherry picked from commit 82505dc351b2f7e37aa395218709b432d83292cd)
