+++ Lib/test/test_future_stmt/__init__.py
@@ -1,0 +1,6 @@
+import os
+from test import support
+
+
+def load_tests(*args):
+    return support.load_package_tests(os.path.dirname(__file__), *args)
