+++ Lib/test/test_future_stmt/badsyntax_future4.py
