+++ Lib/test/test_future_stmt/badsyntax_future6.py
