+++ Lib/test/test_future_stmt/badsyntax_future7.py
