+++ Makefile.pre.in
@@ -2136,6 +2136,7 @@ TESTSUBDIRS=	idlelib/idle_test \
 		test/test_ctypes \
 		test/test_email \
 		test/test_email/data \
+		test/test_future_stmt \
 		test/test_import \
 		test/test_import/data \
 		test/test_import/data/circular_imports \
